Moved generating the emails into being a class method
There is a pattern generating the emails, so made a method to formalise
that and generate emails in the same way every time for us.

Made it a class method instead of an instance method because we need to
be able to use it without there already being an instance.  We know this
because we use it in the constructor.